introduced newer paths and with newer methods. Updated utils.js as well
