+++ Server/middleware/validateUserInput.js
@@ -1,5 +1,4 @@
 const { body, validationResult } = require('express-validator');
-
 const validateRegisterInput = [
     // Sanitize input fields
     body('name').trim().escape(),
@@ -12,6 +11,9 @@ const validateRegisterInput = [
     body('roles').notEmpty().withMessage('Roles is required.'),
     body('phone').isMobilePhone().withMessage('Invalid phone number.'),
     body('email').isEmail().withMessage('Invalid email address.'),
+    body('username').notEmpty().withMessage('Username is required.'),
+    body('username').isLength({ min: 3, max: 20 }).withMessage('Username must be between 3 and 20 characters.'),
+    body('username').isAlphanumeric().withMessage('Username must contain only letters and numbers.'),
     body('password').notEmpty().withMessage('Password is required.'),
     body('password')
         .isLength({ min: 8 })
@@ -23,16 +25,11 @@ const validateRegisterInput = [
         .matches(/[!@#$%^&*(),.?":{}|<>]/)
         .withMessage('Password must contain at least one special character'),
 ];
-
 const validateLoginInput = [
-    body('email').trim().normalizeEmail(),
-    body('email').isEmail().withMessage('Invalid email address.'),
-
+    body('email').trim().normalizeEmail().optional(),
+    body('email').if(body('email').exists()).isEmail().withMessage('Invalid email address.'),
     body('password').notEmpty().withMessage('Password is required.'),
-]
-
-
-
+];
 const handleValidationErrors = (req, res, next) => {
     const errors = validationResult(req);
     if (!errors.isEmpty()) {
@@ -49,6 +46,5 @@ const handleValidationErrors = (req, res, next) => {
     }
     next(); // Proceed to the next middleware/route handler if validation passes
 };
-
 module.exports = {validateRegisterInput, validateLoginInput, handleValidationErrors};
